changing model folder structure
