+++ Dockerfile
@@ -1,0 +1,6 @@
+FROM python:3.7-alpine
+
+COPY ./ /app
+WORKDIR /app
+RUN pip install -r requirements.txt 
+CMD [ "python","main.py"]
