--- app/main.py
@@ -1,8 +1,0 @@
-from fastapi import FastAPI
-
-app=FastAPI()
-
-
-if __name__=="main":
-    predictperson()
-
